# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * JARVIS personal context and automation API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface ChatInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
}

export interface ChatSource {
  type: string;
  title: string;
  excerpt: string;
}

export interface ChatResponse {
  reply: string;
  sources: ChatSource[];
}

export interface Memory {
  id: number;
  content: string;
  category: string;
  importance: number;
  source: string;
  createdAt: string;
  updatedAt: string;
}

export interface MemoryInput {
  /** @minLength 1 */
  content: string;
  category?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  importance?: number;
}

export interface MemoryUpdate {
  /** @minLength 1 */
  content?: string;
  category?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  importance?: number;
}

export interface Document {
  id: number;
  filename: string;
  title: string;
  fileSize: number;
  processingStatus: string;
  chunkCount: number;
  createdAt: string;
}

export type DocumentDetail = Document & {
  textPreview: string;
};

export interface DocumentInput {
  /** @minLength 1 */
  filename: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  /** @minimum 0 */
  fileSize?: number;
}

export interface UploadDocumentInput {
  file: Blob;
  title?: string;
}

export interface DocumentQueryInput {
  /** @minLength 1 */
  query: string;
  /** @minimum 1 */
  documentId?: number;
}

export interface DocumentQuerySource {
  documentId: number;
  title: string;
  excerpt: string;
}

export interface DocumentQueryResponse {
  answer: string;
  sources: DocumentQuerySource[];
}

export interface Goal {
  id: number;
  title: string;
  description: string;
  category: string;
  priority: string;
  status: string;
  progress: number;
  /** @nullable */
  deadline: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Milestone {
  id: number;
  goalId: number;
  title: string;
  description: string;
  status: string;
  progress: number;
  /** @nullable */
  dueDate: string | null;
  createdAt: string;
}

export type GoalDetail = Goal & {
  milestones: Milestone[];
};

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  category?: string;
  priority?: string;
  /** @nullable */
  deadline?: string | null;
}

export interface GoalUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: string;
  priority?: string;
  status?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  /** @nullable */
  deadline?: string | null;
}

export interface MilestoneInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface MilestoneUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  /** @nullable */
  dueDate?: string | null;
}

export interface Meeting {
  id: number;
  title: string;
  transcript: string;
  summary: string;
  decisions: string[];
  actionItems: string[];
  participants: string[];
  /** @nullable */
  meetingDate: string | null;
  createdAt: string;
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 20 */
  transcript: string;
  /** @nullable */
  meetingDate?: string | null;
}

export interface Reminder {
  id: number;
  title: string;
  description: string;
  dueAt: string;
  status: string;
  source: string;
}

export interface ReminderInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  dueAt: string;
}

export interface ReminderUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  dueAt?: string;
  status?: string;
}

export interface Dashboard {
  greeting: string;
  focus: string;
  activeGoals: number;
  completedGoals: number;
  upcomingReminders: Reminder[];
  recentMemories: Memory[];
  recentDocuments: Document[];
  recentMeetings: Meeting[];
  xp: number;
  level: number;
  streak: number;
}

export interface Growth {
  xp: number;
  level: number;
  nextLevelXp: number;
  streak: number;
  activeGoals: number;
  completedGoals: number;
  completionRate: number;
  completedTasks: number;
  memoriesCreated: number;
  documentsUploaded: number;
  meetingsProcessed: number;
  skills: string[];
}

export interface TimelineEvent {
  id: string;
  kind: string;
  title: string;
  description: string;
  occurredAt: string;
}

export interface KnowledgeResult {
  kind: string;
  title: string;
  excerpt: string;
  occurredAt: string;
}

export type SearchKnowledgeParams = {
/**
 * @minLength 1
 */
q: string;
};