# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * JARVIS personal context and automation API
 * OpenAPI spec version: 0.2.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
}